画像の削除処理実装完了

+++ ChatApp/app.py
@@ -103,8 +103,9 @@ def delete_img(category, id):
         img_path = dbConnect.getUserById(id)['profile_img']
     elif category == 'group':
         img_path = dbConnect.getGroupById(id)['group_img']
-    os.remove(f'static/{img_path}')
-    print(f'img_path>>{img_path}')
+    if os.path.exists(f'static/{img_path}'):
+        os.remove(f'static/{img_path}')
+        print(f'img_path>>{img_path}')
     return 'delete_img OK'
 
 
@@ -116,86 +117,9 @@ def delete_users_img(uid):
     os.remove(f'static/{profile_img}')
     users_group_images = dbConnect.getGroupAllByCreateUer(uid)
     for group_image in users_group_images:
+        if os.path.exists(f'static/{group_image}'):
+            os.remove(f'static/{group_image}')
         print(f'グループ写真を削除>>{group_image}')
-        os.remove(f'static/{group_image}')
-    return 'delete_img OK'
-
-
-# =========================
-# 画像関連
-# =========================
-
-# 拡張子を確認し、ファイル名を付ける
-def generate_filename(file):
-    origin_filename = file.filename
-    ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg'}
-    extension = origin_filename.rsplit('.', 1)[-1].lower()
-    if extension in ALLOWED_EXTENSIONS:
-        filename = secure_filename(origin_filename)
-        timestamp = int(time.time())
-        return f"{timestamp}_{filename}"
-    else:
-        flash('画像を選んでね', 'flash ng')
-        return 'Not img'
-
-
-# プロフィール画像をフォルダに保存する
-def profile_img_save(template):
-    # 画像の保存先ディレクトリ
-    PROFILE_IMG_FOLDER = 'static/img/profile_img'
-    app.config[PROFILE_IMG_FOLDER] = PROFILE_IMG_FOLDER
-    profile_img = None
-    if 'profile_img' in request.files:
-        file = request.files['profile_img']
-        if file:
-            filename = generate_filename(file)
-            # 拡張子が画像でない場合　
-            if filename == 'Not img':
-                return redirect(url_for(template))
-            file.save(os.path.join(app.config[PROFILE_IMG_FOLDER], filename))
-            profile_img = f"img/profile_img/{filename}"
-    return profile_img
-
-
-# グループ画像をフォルダに保存する
-def group_img_save(template):
-    # 画像の保存先ディレクトリ
-    GROUP_IMG_FOLDER = 'static/img/group_img'
-    app.config[GROUP_IMG_FOLDER] = GROUP_IMG_FOLDER
-    group_img = None
-    if 'group_img' in request.files:
-        file = request.files['group_img']
-        if file:
-            filename = generate_filename(file)
-            # 拡張子が画像でない場合　
-            if filename == 'Not img':
-                return redirect(url_for(template))
-            file.save(os.path.join(app.config[GROUP_IMG_FOLDER], filename))
-            group_img = f"img/group_img/{filename}"
-    return group_img
-
-
-# DBの画像パスを更新するときに、フォルダにある古い画像を削除する
-def delete_img(category, id):
-    if category == 'profile':
-        img_path = dbConnect.getUserById(id)['profile_img']
-    elif category == 'group':
-        img_path = dbConnect.getGroupById(id)['group_img']
-    os.remove(f'static/{img_path}')
-    print(f'img_path>>{img_path}')
-    return 'delete_img OK'
-
-
-# 退会した人の画像をフォルダから削除する
-def delete_users_img(uid):
-    profile_img = dbConnect.getUserById(uid)['profile_img']
-    print(f'プロフィール写真を削除>>{profile_img}')
-    print(f'static/{profile_img}')
-    os.remove(f'static/{profile_img}')
-    users_group_images = dbConnect.getGroupAllByCreateUer(uid)
-    for group_image in users_group_images:
-        print(f'グループ写真を削除>>{group_image}')
-        os.remove(f'static/{group_image}')
     return 'delete_img OK'
 
 
@@ -320,6 +244,7 @@ def withdraw_account():
     DB_user = session_check()
     if DB_user is not None:
         uid = DB_user['uid']
+        delete_users_img(uid)
         dbConnect.deactivateUser(uid)
         session.clear()
         flash('退会処理が完了しました。またいつでも遊びにきてね！', 'flash ok')
@@ -378,12 +303,15 @@ def update():
         password = hashlib.sha256(password1.encode('utf-8')).hexdigest()
     else:
         password = DB_user["password"]
+
     profile_img = profile_img_save(template='/update_profile')
+    uid = DB_user['uid']
     if profile_img is None:
         user = DB_user
         profile_img = user['profile_img']
-
-    uid = DB_user['uid']
+    else:
+        # 更新があれば前の画像を削除
+        delete_img('profile', uid)
     dbConnect.updateUser(name, email, password, profile_img, uid)
     return redirect(url_for("home"))
 
@@ -510,6 +438,7 @@ def update_chat_group():
         chat_group = dbConnect.getGroupById(cid)
         if chat_group["uid"] != uid:
             flash('チャットグループは作った人だけが削除できるよ！', 'flash ng')
+        delete_img('group', cid)
         dbConnect.deleteGroup(cid)
     return redirect('/groups')
